requirements.txt
# =============================================================================
# Core Data
# =============================================================================
pandas>=2.0
numpy>=1.26
pyarrow>=15.0
openpyxl>=3.1
sqlalchemy>=2.0

# =============================================================================
# Validation and Profiling
# =============================================================================
great-expectations>=0.18
pandera>=0.19
ydata-profiling>=4.6
jsonschema>=4.20
soda-core>=3.3

# =============================================================================
# NLP and Text
# =============================================================================
spacy>=3.7
pymupdf>=1.24
langdetect>=1.0.9
transformers>=4.40

# =============================================================================
# Audio
# =============================================================================
openai-whisper
librosa>=0.10
pydub>=0.25

# =============================================================================
# Image and Video
# =============================================================================
opencv-python>=4.9
Pillow>=10.0
imagehash>=4.3
moviepy>=1.0

# =============================================================================
# API and Serving
# =============================================================================
flask>=3.0
flask-cors>=4.0
requests>=2.31
httpx>=0.27

# =============================================================================
# CI/CD and Infrastructure
# =============================================================================
dvc>=3.0
dvc-s3>=3.0

# =============================================================================
# Alerts and Reporting
# =============================================================================
slack-sdk>=3.27
boto3>=1.34

# =============================================================================
# Testing
# =============================================================================
pytest>=8.0
pytest-cov>=5.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.66
loguru>=0.7
